Integrate Prodigal CDS listing
