+++ analyse_seq.py
@@ -3,8 +3,71 @@ import argparse
 import os
 import subprocess
 from typing import Dict, List
-
 from ete3 import Tree
+
+
+def run_prodigal(fasta: str, prefix: str, mode: str = "meta") -> tuple[str, str]:
+    """Run prodigal and return paths to GFF and protein FASTA."""
+    gff = f"{prefix}.gff"
+    faa = f"{prefix}.faa"
+    subprocess.run(
+        [
+            "prodigal",
+            "-i",
+            fasta,
+            "-p",
+            mode,
+            "-a",
+            faa,
+            "-f",
+            "gff",
+            "-o",
+            gff,
+            "-q",
+        ],
+        check=True,
+    )
+    return gff, faa
+
+
+def parse_gff(gff: str) -> List[Dict[str, str]]:
+    """Parse Prodigal GFF and return CDS info."""
+    cds = []
+    with open(gff) as fh:
+        for line in fh:
+            if line.startswith("#"):
+                continue
+            parts = line.strip().split("\t")
+            if len(parts) < 9:
+                continue
+            seqid, _source, type_, start, end, _score, strand, _phase, attrs = parts
+            if type_ != "CDS":
+                continue
+            attr_dict: Dict[str, str] = {}
+            for item in attrs.split(";"):
+                if "=" in item:
+                    key, value = item.split("=", 1)
+                    attr_dict[key] = value
+            cds.append(
+                {
+                    "seqid": seqid,
+                    "start": start,
+                    "end": end,
+                    "strand": strand,
+                    "id": attr_dict.get("ID", ""),
+                }
+            )
+    return cds
+
+
+def summarize_cds(cds: List[Dict[str, str]]) -> None:
+    """Print a short summary of CDS."""
+    print(f"{len(cds)} CDS predicted")
+    strand_counts: Dict[str, int] = {}
+    for c in cds:
+        strand_counts[c["strand"]] = strand_counts.get(c["strand"], 0) + 1
+    for strand, count in strand_counts.items():
+        print(f"{count} CDS on strand {strand}")
 
 
 def compute_gc(fasta: str) -> float:
@@ -210,6 +273,22 @@ def main():
     parser.add_argument(
         "--tmpdir", default="tmp", help="Répertoire pour les fichiers temporaires"
     )
+    parser.add_argument(
+        "--list-cds",
+        action="store_true",
+        help="Lancer Prodigal et afficher un résumé des CDS prédits",
+    )
+    parser.add_argument(
+        "--prodigal-prefix",
+        default="prodigal",
+        help="Préfixe des fichiers générés par Prodigal",
+    )
+    parser.add_argument(
+        "--prodigal-mode",
+        choices=["meta", "single"],
+        default="meta",
+        help="Mode de Prodigal (meta ou single)",
+    )
     parser.add_argument("--evalue", type=float, default=1e-5)
     parser.add_argument(
         "--orf-search",
@@ -284,6 +363,18 @@ def main():
             print(msg)
         except RuntimeError as err:
             print(err)
+
+    if args.list_cds:
+        prefix = os.path.join(args.tmpdir, args.prodigal_prefix)
+        os.makedirs(os.path.dirname(prefix), exist_ok=True)
+        try:
+            gff, faa = run_prodigal(args.fasta, prefix, args.prodigal_mode)
+            cds = parse_gff(gff)
+            summarize_cds(cds)
+            print(f"CDS annotations written to {gff}")
+            print(f"Protein translations written to {faa}")
+        except Exception as err:
+            print(f"Prodigal failed: {err}")
 
     # Désactivation par défaut de TransposonPSI car obsolète/incompatible
     # if args.transposonpsi:
